rare-disease-mm-ai-bench/requirements.txt

torch>=2.1
torchvision>=0.16
torchaudio>=2.1
transformers>=4.40
datasets>=2.19
scikit-learn>=1.4
pandas>=2.2
numpy>=1.26
tqdm>=4.66
pyyaml>=6.0.1
opencv-python>=4.10
pillow>=10.3
einops>=0.8.0
matplotlib>=3.8
